fix(autoware_path_generator): fix bugprone-narrowing-conversions warnings (#921)
* fix(autoware_path_generator): fix bugprone-narrowing-conversions warnings


* style(pre-commit): autofix

---------

+++ planning/autoware_path_generator/src/utils.cpp
@@ -291,7 +291,7 @@ std::optional<double> get_first_self_intersection_arc_length(
     }
 
     const auto current_segment = lanelet::BasicSegment2d{line_string.at(i), line_string.at(i + 1)};
-    s += lanelet::geometry::length(current_segment);
+    s += static_cast<double>(lanelet::geometry::length(current_segment));
 
     lanelet::BasicPoints2d self_intersections{};
     for (size_t j = i + 1; j < line_string.size() - 1; ++j) {
@@ -465,7 +465,8 @@ double get_arc_length_on_path(
 
   for (auto [it, s] = std::make_tuple(lanelet_sequence.begin(), 0.); it != lanelet_sequence.end();
        ++it) {
-    const double centerline_length = lanelet::geometry::length(it->centerline2d());
+    const double centerline_length =
+      static_cast<double>(lanelet::geometry::length(it->centerline2d()));
     if (s + centerline_length < s_centerline) {
       s += centerline_length;
       continue;
@@ -578,9 +579,12 @@ PathRange<double> get_arc_length_on_bounds(
   auto s_right = 0.;
 
   for (auto it = lanelet_sequence.begin(); it != lanelet_sequence.end(); ++it) {
-    const double centerline_length = lanelet::geometry::length(it->centerline2d());
-    const double left_bound_length = lanelet::geometry::length(it->leftBound2d());
-    const double right_bound_length = lanelet::geometry::length(it->rightBound2d());
+    const double centerline_length =
+      static_cast<double>(lanelet::geometry::length(it->centerline2d()));
+    const double left_bound_length =
+      static_cast<double>(lanelet::geometry::length(it->leftBound2d()));
+    const double right_bound_length =
+      static_cast<double>(lanelet::geometry::length(it->rightBound2d()));
 
     if (s + centerline_length < s_centerline) {
       s += centerline_length;
@@ -633,9 +637,11 @@ PathRange<std::optional<double>> get_arc_length_on_centerline(
       break;
     }
 
-    const double centerline_length = lanelet::geometry::length2d(*it);
-    const double left_bound_length = lanelet::geometry::length(it->leftBound2d());
-    const double right_bound_length = lanelet::geometry::length(it->rightBound2d());
+    const double centerline_length = static_cast<double>(lanelet::geometry::length2d(*it));
+    const double left_bound_length =
+      static_cast<double>(lanelet::geometry::length(it->leftBound2d()));
+    const double right_bound_length =
+      static_cast<double>(lanelet::geometry::length(it->rightBound2d()));
 
     if (!is_left_done && s_left + left_bound_length > s_left_bound) {
       s_left_centerline = s + lanelet::geometry::toArcCoordinates(
